[plugin-docker] fix CPU/Memory metrics on Docker hosts uses cgroup2
Fix the issue mackerel-plugin-docker outputs always zero value for CPU/Memory metrics.
It is because recent docker daemon has been changed to cgroup2 by default and usage of cgroup2 varies the output of Docker Stats API.
So this commit fixes the usage of Stats API output to adapt cgroup2 hosts.
See #885 for detail of this problem.